refactor(validation): fold abstractions into core as RequestValidation
Maintainer feedback on #4183: move the provider-agnostic validation
abstractions out of the separate Paramore.Brighter.Validation package into the
core Paramore.Brighter assembly under RequestValidation/, mirroring the built-in
RequestLogging/UseInbox attribute+handler pattern. Providers stay separate
packages.

- Rename [ValidateQuery]/[ValidateQueryAsync] -> [ValidateRequest]/
  [ValidateRequestAsync]; Brighter has requests, not queries.
- Rename the error type to RequestValidationError to avoid clashing with the
  pre-existing pipeline-validation Paramore.Brighter.ValidationError.
- Use file-scoped namespaces to match the current core style.
- Point the FluentValidation provider at the core directly; delete the
  standalone abstractions project and its slnx entry.
- Drop the package-level <Authors> override (repo-wide Authors apply); add the
  contributor to CONTRIBUTORS.md.
- Update ADR 0063 to record the decisions and resolved review questions.

Contributes to #4175
