+++ tests/Paramore.Brighter.Validation.FluentValidation.Tests/When_an_invalid_request_has_multiple_errors_should_report_them_all.cs
@@ -24,6 +24,7 @@ THE SOFTWARE. */
 
 using System.Linq;
 using Microsoft.Extensions.DependencyInjection;
+using Paramore.Brighter.RequestValidation;
 using Paramore.Brighter.Validation.FluentValidation.Tests.TestDoubles;
 using Xunit;
 using global::FluentValidation;
