+++ tests/Paramore.Brighter.Validation.FluentValidation.Tests/When_an_invalid_request_is_sent_async_should_throw_request_validation_exception.cs
@@ -23,6 +23,7 @@ THE SOFTWARE. */
 #endregion
 
 using System.Threading.Tasks;
+using Paramore.Brighter.RequestValidation;
 using Paramore.Brighter.Validation.FluentValidation.Tests.TestDoubles;
 using Xunit;
