+++ tests/Paramore.Brighter.Validation.FluentValidation.Tests/When_an_invalid_request_is_sent_should_throw_request_validation_exception.cs
@@ -22,6 +22,7 @@ THE SOFTWARE. */
 
 #endregion
 
+using Paramore.Brighter.RequestValidation;
 using Paramore.Brighter.Validation.FluentValidation.Tests.TestDoubles;
 using Xunit;
