+++ tests/Paramore.Brighter.Validation.FluentValidation.Tests/When_the_handler_is_used_concurrently_should_validate_each_request.cs
@@ -24,6 +24,7 @@ THE SOFTWARE. */
 
 using System.Threading.Tasks;
 using Microsoft.Extensions.DependencyInjection;
+using Paramore.Brighter.RequestValidation;
 using Paramore.Brighter.Validation.FluentValidation.Tests.TestDoubles;
 using Xunit;
 using global::FluentValidation;
